Interface and model tinkered with. loadTemperatureData method written. Some errors with JSON response from server

+++ app/src/main/java/com/example/greenhousecompanion/api/model/TemperatureModel.kt
@@ -1,7 +1,5 @@
 package com.example.greenhousecompanion.api.model
 
 data class TemperatureData(
-    val temperature: Double
+    val temperature: Float
 )
-
-
